First Release for Testing (#1)
* feat: update packages

* feat: update packages

* feat: add store info

+++ src/stores/stores.service.ts
@@ -16,6 +16,7 @@ import { InvitationDto } from './dto/invitation.dto';
 import { UserRepository } from 'src/users/repositories/users.repository';
 import { MemberRole } from 'src/generated/prisma/enums';
 import { sendInvitationEmail } from 'src/helper/function/resend-email';
+import { ProductRepository } from 'src/products/repositories/products.repository';
 
 @Injectable()
 export class StoresService {
@@ -23,6 +24,7 @@ export class StoresService {
     private readonly userRepo: UserRepository,
     private readonly storeRepo: StoreRepository,
     private readonly memberRepo: MemberRepository,
+    private readonly productRepo: ProductRepository,
   ) {}
 
   findAll(query: QueryStoreDto, userId: number) {
@@ -43,17 +45,67 @@ export class StoresService {
     return { success: true, ...createStore, members: addingOwner };
   }
 
-  updateStore(updateDto: UpdateStoreDto, storeId: number) {
+  async updateStore(
+    updateDto: UpdateStoreDto,
+    storeId: number,
+    userId: number,
+  ) {
+    const store = await this.storeRepo.findOne(storeId);
+    if (!store) {
+      throw new NotFoundException({
+        message: 'Store not found',
+        errorCode: ErrorCode.STORE_NOT_FOUND,
+        statusCode: HttpStatus.NOT_FOUND,
+      });
+    }
+    const member = await this.memberRepo.findOne(storeId, userId);
+    if (!member) {
+      throw new NotFoundException({
+        message: 'Member not found',
+        errorCode: ErrorCode.MEMBER_NOT_FOUND,
+        statusCode: HttpStatus.NOT_FOUND,
+      });
+    }
+    if (member.role !== 'OWNER' && member.role !== 'MANAGER') {
+      throw new ForbiddenException({
+        message: 'You are not authorized to update this store',
+        errorCode: ErrorCode.INSUFFICIENT_PERMISSIONS,
+        statusCode: HttpStatus.FORBIDDEN,
+      });
+    }
     return this.storeRepo.updateStore(updateDto, storeId);
   }
 
-  async deleteStore(storeId: number) {
+  async deleteStore(storeId: number, userId: number) {
+    const store = await this.storeRepo.findOne(storeId);
+    if (!store) {
+      throw new NotFoundException({
+        message: 'Store not found',
+        errorCode: ErrorCode.STORE_NOT_FOUND,
+        statusCode: HttpStatus.NOT_FOUND,
+      });
+    }
+    const member = await this.memberRepo.findOne(storeId, userId);
+    if (!member) {
+      throw new NotFoundException({
+        message: 'Member not found',
+        errorCode: ErrorCode.MEMBER_NOT_FOUND,
+        statusCode: HttpStatus.NOT_FOUND,
+      });
+    }
+    if (member.role !== 'OWNER') {
+      throw new ForbiddenException({
+        message: 'You are not authorized to delete this store',
+        errorCode: ErrorCode.INSUFFICIENT_PERMISSIONS,
+        statusCode: HttpStatus.FORBIDDEN,
+      });
+    }
     await this.storeRepo.deleteStore(storeId);
     return { success: true, message: 'Store deleted successfully' };
   }
 
   // Invitation logic
-  async createInviteLink(storeId: number) {
+  async createInviteLink(storeId: number, userId: number) {
     const { token } = generateToken();
     const store = await this.storeRepo.findOne(storeId);
     if (!store) {
@@ -63,18 +115,68 @@ export class StoresService {
         statusCode: HttpStatus.NOT_FOUND,
       });
     }
+    const member = await this.memberRepo.findOne(storeId, userId);
+    if (!member) {
+      throw new NotFoundException({
+        message: 'Member not found',
+        errorCode: ErrorCode.MEMBER_NOT_FOUND,
+        statusCode: HttpStatus.NOT_FOUND,
+      });
+    }
+    if (member.role !== 'OWNER' && member.role !== 'MANAGER') {
+      throw new ForbiddenException({
+        message: 'You are not authorized to create invite link',
+        errorCode: ErrorCode.INSUFFICIENT_PERMISSIONS,
+        statusCode: HttpStatus.FORBIDDEN,
+      });
+    }
     await this.storeRepo.createInviteLink(token, storeId);
     return { inviteCode: token };
   }
 
-  async getInviteLink(storeId: number) {
+  async getInviteLink(storeId: number, userId: number) {
     const store = await this.storeRepo.findOne(storeId);
+    const member = await this.memberRepo.findOne(storeId, userId);
+    if (!member) {
+      throw new NotFoundException({
+        message: 'Member not found',
+        errorCode: ErrorCode.MEMBER_NOT_FOUND,
+        statusCode: HttpStatus.NOT_FOUND,
+      });
+    }
+    if (member.role !== 'OWNER' && member.role !== 'MANAGER') {
+      throw new ForbiddenException({
+        message: 'You are not authorized to get invite link',
+        errorCode: ErrorCode.INSUFFICIENT_PERMISSIONS,
+        statusCode: HttpStatus.FORBIDDEN,
+      });
+    }
     return { inviteCode: store?.inviteCode };
   }
 
-  async createInvitation(invitationDto: InvitationDto, storeId: number) {
+  async createInvitation(
+    invitationDto: InvitationDto,
+    storeId: number,
+    userId: number,
+  ) {
+    const member = await this.memberRepo.findOne(storeId, userId);
     const user = await this.userRepo.findByEmail(invitationDto.email);
-    let token = await this.getInviteLink(storeId);
+    let token = await this.getInviteLink(storeId, userId);
+
+    if (!member) {
+      throw new NotFoundException({
+        message: 'Member not found',
+        errorCode: ErrorCode.MEMBER_NOT_FOUND,
+        statusCode: HttpStatus.NOT_FOUND,
+      });
+    }
+    if (member.role !== 'OWNER' && member.role !== 'MANAGER') {
+      throw new ForbiddenException({
+        message: 'You are not authorized to get invite link',
+        errorCode: ErrorCode.INSUFFICIENT_PERMISSIONS,
+        statusCode: HttpStatus.FORBIDDEN,
+      });
+    }
 
     if (!user) {
       throw new NotFoundException({
@@ -83,8 +185,6 @@ export class StoresService {
         statusCode: HttpStatus.NOT_FOUND,
       });
     }
-
-    const member = await this.memberRepo.findOne(storeId, user.userId);
 
     if (member) {
       throw new ConflictException({
@@ -95,7 +195,7 @@ export class StoresService {
     }
 
     if (!token.inviteCode) {
-      token = await this.createInviteLink(storeId);
+      token = await this.createInviteLink(storeId, userId);
     }
     if (!token.inviteCode) {
       throw new NotFoundException({
@@ -115,7 +215,22 @@ export class StoresService {
     return { success: true, invitation: created };
   }
 
-  async getInvitations(storeId: number) {
+  async getInvitations(storeId: number, userId: number) {
+    const member = await this.memberRepo.findOne(storeId, userId);
+    if (!member) {
+      throw new NotFoundException({
+        message: 'Member not found',
+        errorCode: ErrorCode.MEMBER_NOT_FOUND,
+        statusCode: HttpStatus.NOT_FOUND,
+      });
+    }
+    if (member.role !== 'OWNER' && member.role !== 'MANAGER') {
+      throw new ForbiddenException({
+        message: 'You are not authorized to get invite link',
+        errorCode: ErrorCode.INSUFFICIENT_PERMISSIONS,
+        statusCode: HttpStatus.FORBIDDEN,
+      });
+    }
     const invitations = await this.storeRepo.getInvitations(storeId);
     return invitations.map(({ memberId, ...invitation }) => {
       return {
@@ -155,5 +270,30 @@ export class StoresService {
       });
     }
     return this.memberRepo.deleteMember(storeId, targetMember.userId);
+  }
+
+  // Get Store Info
+  async getStoreInfo(storeId: number, userId: number) {
+    const store = await this.storeRepo.findOne(storeId);
+    if (!store) {
+      throw new NotFoundException({
+        message: 'Store not found',
+        errorCode: ErrorCode.STORE_NOT_FOUND,
+        statusCode: HttpStatus.NOT_FOUND,
+      });
+    }
+    const member = await this.memberRepo.findOne(storeId, userId);
+    if (!member) {
+      throw new ForbiddenException({
+        message: 'You are not a member of this store',
+        errorCode: ErrorCode.INSUFFICIENT_PERMISSIONS,
+        statusCode: HttpStatus.FORBIDDEN,
+      });
+    }
+
+    const memberCount = await this.memberRepo.getMemberCount(storeId);
+    const productsCount = await this.productRepo.getProductCount(storeId);
+
+    return { storeId: store.storeId, memberCount, productsCount };
   }
 }
